LayoutTests/imported/w3c/web-platform-tests/css/css-syntax/serialize-escape-identifiers-expected.txt


FAIL Rules must be the same after serialization round-trip, even with escaped characters assert_equals: expected "@import url(\"abc\") layer(\\{\\});\n@font-feature-values \"abc{}oops\" { }\n@font-palette-values --abc{}oops { }\n@keyframes abc{}oops { \n}\n@layer abc\\;oops\\!;" but got "@import url(\"abc\") layer(\\{\\});\n@font-feature-values \"abc{}oops\" { }\n@font-palette-values --abc { }\noops { }\n@keyframes abc { \n}\noops { }\n@layer abc\\;oops\\!;"
